Revert "Refactor: change folder layout"
