+++ src/app/hooks/useHash.ts
