+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import LandingPage from "../components/landingPage/LandingPage";
-import MeetUs from "../components/meetUs/MeetUs";
-import Navbar from "../components/navBar/Navbar";
+import LandingPage from "./components/landingPage/LandingPage";
+import MeetUs from "./components/meetUs/MeetUs";
+import Navbar from "./components/navBar/Navbar";
 import { useGetFoodsQuery } from "./getFoods.rq.generated";
 import styles from "./page.module.scss";
